Convert many non-ASCII characters with ASCII
This is a must for factor levels and other information that users will type
in code. It is also done in the description of different codes if the
non-ASCII character is easily confused with an ASCII one.
Greek letters are left as-is in descriptive fields,
because analysts should understand them but not need to type them.